applied jekyll hacker theme for basic use
The gem was added into the Gemfile and config updated to use it.
Existing example documents updated to use new style as their `page` and
`home` layouts do not exist in the selected theme.
The index page was also changed to display a list (of 1 currently!)
posts.

It needs to be paginated and the minimum information provided in a post
needs defining, but this is enough of a change to test the daily update
on the server which should pull the master branch on github.