removed global permalink
The permalink was set /:title which meant the posts were being created
in the root of the web site and I wanted them to be created in a
year/month/day structure, although as i write this a year/month
structure would be better as I am not likely to create more than one
post in a day (or a week for that matter).

I may change this after I deploy and test on the real server.