dealt with SASS deprecation of @import in a poor way
I read the SASS site and other sites and through experiment it is clear
the information provided about @use and @forward do not work for me.  I
ended up putting it all in one file because the ability to re-use
anything, and one of the main reasons for SASS were impossible to get
working for me.

Simply replacing @import with @use doesn't work.  I have a partial, and
if I change it into a proper file, it doesn't work.  If I @use each file
in ever file it doesn't work.

I'm at a total loss and will need to find a book to teach me how new SAS
with @USE and @FORWARD and partials work.